+++ lib/arxiv/downloader.rb
@@ -18,6 +18,7 @@ require_relative 'downloader/metadata/json'
 require_relative 'downloader/metadata/bibtex'
 require_relative 'downloader/metadata/markdown'
 require_relative 'downloader/archive'
+require_relative 'downloader/cli'
 
 module Arxiv
   module Downloader